requirements.txt
colorama==0.4.6
easyocr==1.7.1
Flask==3.0.3
fuzzywuzzy==0.18.0
Pillow==10.4.0
Requests==2.32.3
torch==2.4.0+cu118
